+++ devDungeon/src/level/devlevel/riddleHandler/TorchRiddleRiddleHandler.java
@@ -4,6 +4,7 @@ import components.TorchComponent;
 import contrib.components.SignComponent;
 import contrib.entities.SignFactory;
 import contrib.hud.DialogUtils;
+import contrib.utils.components.skill.projectileSkill.BurningFireballSkill;
 import core.Entity;
 import core.Game;
 import core.components.DrawComponent;
@@ -14,7 +15,6 @@ import core.level.utils.Coordinate;
 import core.level.utils.LevelUtils;
 import core.utils.Point;
 import core.utils.components.MissingComponentException;
-import entities.BurningFireballSkill;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.IntStream;
